# SampleNetCore/src/SampleNetCore/Swagger/sample.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "REST API Documentation",
    "description": "API Sample",
    "termsOfService": "None"
  },
  "basePath": "/",
  "paths": {
    "/api/Values": {
      "get": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": { "type": "string" }
            }
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Values/{id}": {
      "get": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesByIdGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "string" }
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesByIdPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      },
      "delete": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesByIdDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/Vehicles": {
      "get": {
        "tags": [ "Vehicle" ],
        "operationId": "VehiclesGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Vehicle" }
            }
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
    "Vehicle": {
      "type": "object",
      "properties": {
        "vehicle_id": {
          "format": "int32",
          "type": "integer"
        },
        "description": { "type": "string" },
        "display_name": { "type": "string" }
      }
    }
  },
  "securityDefinitions": {}
}